Reorganise module
